Fix for IE8 calendar/time collapse
In IE8, if you call enable on a picker that's not disabled, the click handler code gets called twice on click. This causes the calendar to collapse and expand immediately making it appear that nothing is working. Making sure that an picker is disable before enabling (and vise versa) prevents these shenanigans and doesn't impact behavior on other browsers.