Fixed a bug when rendering months with daylight saving time
Within the fillDate method, on line 386, there was a bug with the while:

    while (prevMonth.isBefore(nextMonth)) {

When a day has less than 24 hours (23 because of daylight saving), the
following call will never reach the next day:

    prevMonth.add(1, "d");

because prevMonth is at 00:00:00 hours.

To fix this, I check if the date moved and repeat the add-day call if
necessary.