Use setDate instead of setValue when useCurrent enabled
This allows dp.change to be triggered, since setting the date to current is
actually a change and should be treated as so. This is useful for situations
where one needs to bind an event handler to the dp.change event. In case the
user actually wants to select the current date (which is preselected on show if using
useCurrent is true), the event is not triggered thus the event handler is not
fired properly, although intended.