update composer.json to remove component/boostrap
reason: 

twbs/bootstrap has a composer.json already, and includes all its components, many of which are needed if you want to run a post-install build process to combine and minify only a subset. This lets the resulting js/css files be much smaller.

component/bootstrap has the component .less files but not the .js files, and robloach/component-installer copies a lot of boostraps files into a <webroot>/component/ directory which is annoying and can't be turned off

removing the dependency allows us the choice. it is fairly obvious from the name of your project that bootstrap is a requirement