Adding a setCalendar to set the month you see
I used the setMinDate but when I opened the calendar where the date was set it still opened on today's month (not the month I just set with setMinDate).

I'm not sure this is the best solution, or is even a valid solution. But for it seems like it works, and I thought let's share it :)

Please let me know what you guy's think!