Add ability to disable and enable inline datepicker
Disabling the datepicker was causing it to be hidden which prevented it
from being enabled.  I added functionality so that when disabled for
inline datepickers all events are removed and the control is basically
set to disabled, re-enabling the datepicker adds the events back and
updates the control.
I also fixed an issue with the css when disabling elements, this was an
issue for the forward and back buttons on the control when they were
disabled but still had the hover effect.  This was fixed by tweaking the
css and the specificity.