Fix bug in disabledTimeIntervals
The option was disabling full hours even if there was selectable minute intervals in those. Ex:

    disabledTimeIntervals: [
        [moment('2015-09-25 17:50'), moment('2015-09-25 18:50')],
    ]

The above configuration would disable hour 18, even though the minutes between 18:50 and 18:55 should be selectable. The interval now is properly disabled.

It's important to note the disabled interval has to be open. Otherwise scheduling applications of this feature would be challenging. This means an interval 18:00 to 19:00 would allow the user to select the times 18:00 and 19:00 but nothing in between.