Support ISO 8601 calendar
Add support for ISO 8601 calendar format in the 'days' view,
enabled when the `isoWeeks` option is set to true.

In ISO 8601 calendar, the week should start on the Monday and
the first week of the year is the week with the first Thursday
of the year in it.

This is an important use case for businesses outside of the US,
especially those operating in Western Europe.