Issue1480 Add option to prevent time adjustments from affecting other date/time elements
Added a noRollTime option (defaults false) that prevents time
adjustments from affecting other date/time elements. e.g. adjust hours
past 23 or below 0 will not affect the date.  Same code is applied to
minutes (not affecting hours) and seconds (not affecting minutes).  In
the case of minutes the stepping option is accounted for.