Fix Right\Left key press in input text
When you type the date and want to use the Right or Left keys to navigate through the date text, the navigation didn't work due to the stopPropagation and preventDefault was run after a key press. After the fix the handler.call function returns true only if the propagation is wanted.