Passing the string "false" to min/maxDate options.
This was a very small change I added to make my life a tiny bit easier when filling these options with php. In my specific use case, it was easier to just pass a string containing "false" than it was to sometimes pass string and sometimes pass a bool. Might come in handy for others?