Fix broken date when pick time is disabled
What:
In 4.17.37 when the time picker was disabled the moment object returned by
the onChange event was a start of day date.
Since 4.17.42 this is no longer the case (#2083).

Why:
4.17.42 fixed a bug and among other things removed a call to startOf('day') on a moment.

How:
Introducing this startOf('day') call when a date is picked: if the time
picker is disabled, startOf('day') is called on the moment to set, else it isn't called.